Updated Validate_password function
Replaced if char == char.lower() with char.isalphaand (char == char.lower()).

And

Replaced if char == char.upper() with char.isalphaand (char == char.upper()).

This is done to ensure that char is a letter first before checking if it is a lowercase/uppercase letter. This is because a symbol for example would return true for both of these statements, as it's its own upper and lower case.