Updated Validate_password function
Errors used to get displayed twice to the user. To fix this I removed the redundant call to Validate_password before the if statement as it gets called again during the if statement.

There was also another issue where the Error window was displayed followed by the Success window. To fix this I had to not return the messagebox error and just call it, and instead return False this ensures the if statement is false and therefore not return the Success window.