Login GUI Updated
I removed the text="Username" and text="Password" from the entry fields, as it I realised that it doesn't act as a placeholder. I then added labels next to each field. I had to switch from .pack() to .grid() so that I could place the labels to the left of each field. This gave me more control over the layout, which makes it easier to organise the GUI into columns. Finally I added show="*" to the password entry field, so that the user's password is hidden as they type for more security.