Login GUI Updated
I changed event.get() to entry().get and that seems to have solved the issue of "AttributeError: 'Event' object has no attribute 'get'". I also noticed that the password field needed special handling, so i added an extra check for the password using the is_password bool variable. The placeholder should be shown normally with no masking, when the user focuses on it the placeholder disappears, so I make sure show="*" is on. If the field is empty I instead use show="", so that the placeholder is visible.