Login GUI Updated
I added a bottom frame, and in it I added a "Don't already have an account?" text followed by a "Sign up." clickable text that's in blue and so that it looks like a hyperlink. I placed both of these labels inside of a new frame so they're easier to position them next to each other. I also added a placeholder function for the sign up link.