Updated Login GUI
Tkinter doesn't open the window centred in the middle of the screen. To fix this I added variables for window_width and window_height and used it along with winfo_screenwidth() and winfo_screenheight() to get the user's screen resolution. I then calculated the x and y coordinates so that the window is position in the centre of the user's screen. I then applied this to self.root.geometry.

I also added self.root.resizable(False, False) so the user cannot resize the window. This prevents the layout from stretching or breaking.

These changes makes the login screen look much closer to my mock-up.