Updated Login GUI
I changed the background of the window to be #1c1c1c, which is the background colour used in my mock-up. I also changed the background of all labels to #1c1c1c to match the background and changed the labels that were black to white. I also replaced the "blue" colour with #001eff, which is the exact one used in my mock-up.