Updated Login GUI
I created a new main_frame which is a container for all of the login components. This makes it easier to control and manage the layout.
By placing all elements inside the main_frame, I could now use place(relx=0.5, rely=0.5, anchor="center") to position the entire frame exactly in the centre of the window, which is much easier to maintain.