Updated Login GUI
I changed the size and font of the "Login" text for the title, "Username" text for the username label, "Password" for the password label, I also made them bold so they look better. I removed the padding between the labels and the entries, as it looks better. I also changed the size and font for the "Don't already have an account" text and the "Sign up." text and made them bold. I also underlined the "Sign up" text, so it looks more like a hyperlink and closer to my mock-up. I also fixed a typo "Dont't" to "Don't".