Updated Login GUI
Instead of error message popups, I instead replaced it with red text that gets displayed below the login button like it was in my mock-up. I did this by making an error_label and positioned it in between the login button and the bottom frame. I also replaced messagebox.showerror() with self.error_label.config(text"...").