Updated Login GUI
I imported customtkinter and used ctk.CTkEntry instead of tk.Entry, as it has a proper placeholder support. This avoids the problem where the placeholder text was treated as user input during validation.

I replaced every tk.Entry with ctk.CTkEntry from the customtkinter library. CTkEntry has a built in placeholder_text, which displays placeholder text visually but it doesn't count as an input, which is a perfect solution for my problem.