Login GUI Error Fix
I got this error: ValueError: ['fg', 'bg'] are not supported arguments because fg and bg weren't supported by customtkinter. I read the documentation and replaced them with text_color and fg_color.

I also changed the width from a charachter-based value (31) to a pixel-based value (250) because this is how customtkinter works.

I also changed the placeholder for username from "Username:" to "Username"