Updated Login GUI
I improved the usability of the password input by adding a real-time password validation feedback. When the user clicks into the password field, a list of password requirements becomes visible underneath the input box. As the user types, each requirement updates instantly and changes colour depending on whether it has been met.

This prevents the user from repeatedly submitting invalid passwords and receiving error messages afterwards. Instead, it guides the user while they are typing, making the interface more user-friendly and reduces frustration. This also makes the user is clear on what the requirements are.

To do this first I created a new frame called "requirements_frame". This stores every label to be displayed. I then created each label in the frame and gave then a default text colour of red symbolising that the requirement isn't met yet. I then also initially removed the frame from the screen. I then created 2 events focus in and focus out. When the user presses into the input field (focuses in) the requirements_frame grid is displayed. When the user clicks off of it (focuses out) the requirements_frame grid is removed. I then created a new procedure to handle the text colour change to green once a requirement has been met. First I iterated over the password length and checked each requirement with each index. If the requirement is met the text of that label is set to green symbolising the requirement has been met.