check_password_requirements procedure fix
I initially iterated over the password length instead of the characters themselves, which led to an error because an integer index rather than a character, so string methods such as isalpha() could not be used. I changed for i in range(len(password))  to for i in password to resolve this issue.